refactor: update version command usage
