+++ packages/webpack-cli/package.json
@@ -41,11 +41,15 @@
     "colorette": "^2.0.14",
     "commander": "^7.0.0",
     "cross-spawn": "^7.0.3",
+    "envinfo": "^7.7.3",
     "fastest-levenshtein": "^1.0.12",
     "import-local": "^3.0.2",
     "interpret": "^2.2.0",
     "rechoir": "^0.7.0",
     "webpack-merge": "^5.7.3"
+  },
+  "devDependencies": {
+    "@types/envinfo": "^7.8.1"
   },
   "peerDependencies": {
     "webpack": "5.x.x"
